bug fix for windows
git-svn-id: http://windowlicker.googlecode.com/svn/trunk@89 ac37430e-0f47-0410-9073-f3f9481b7c33

+++ unit-tests/com/objogate/wl/driver/tests/JFileChooserDriverTest.java
@@ -36,12 +36,13 @@ public class JFileChooserDriverTest extends AbstractComponentDriverTest<JFileCho
 
     private void delete(File dir) {
         File[] files = dir.listFiles();
-        for (File file : files) {
-            if(file.isDirectory())
-                delete(file);
+        if (files != null)
+            for (File file : files) {
+                if (file.isDirectory())
+                    delete(file);
 
-            file.delete();
-        }
+                file.delete();
+            }
     }
 
     public void testPrintOutSubComponents() throws InterruptedException {
